pyproject.toml
[tool.poetry]
name = "weather_api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "3.8"
fastapi = "^0.63.0"
pydantic = "^1.8.1"
requests = "^2.25.1"
python-jose = "^3.2.0"
passlib = "^1.7.4"
python-multipart = "^0.0.5"
bcrypt = "^3.2.0"

[tool.poetry.dev-dependencies]
uvicorn = "^0.13.4"
autopep8 = "^1.5.5"
python-dotenv = "^0.15.0"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"